+++ lib/net_mosq.c
@@ -543,7 +543,7 @@ int _mosquitto_socket_connect(struct mosquitto *mosq, const char *host, uint16_t
 			return MOSQ_ERR_TLS;
 		}
 		SSL_set_ex_data(mosq->ssl, tls_ex_index_mosq, mosq);
-		bio = BIO_new_socket(sock, BIO_NOCLOSE);
+		bio = BIO_new_socket((int)sock, BIO_NOCLOSE);
 		if(!bio){
 			COMPAT_CLOSE(sock);
 			return MOSQ_ERR_TLS;
